Added Visual Studio 2010 Solution and MSVC changes.
All settings are defined in property sheets (release/debug/arch32/arch64/settings) and should be modified from Property Manager.
To build the project you need the latest boost and TFS SDK paths in BoostDir and TFSSDKDir environment variables.



git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@5325 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ vocation.cpp
@@ -255,7 +255,7 @@ uint64_t Vocation::getReqMana(uint32_t magLevel)
 	if(it != cacheMana.end())
 		return it->second;
 
-	uint64_t reqMana = (uint64_t)(400 * pow(manaMultiplier, magLevel - 1));
+	uint64_t reqMana = (uint64_t)(400 * pow(manaMultiplier, (int) magLevel - 1));
 	if(reqMana % 20 < 10)
 		reqMana = reqMana - (reqMana % 20);
 	else
